little changes

+++ src/main/java/com/baran/charactersheetcreator/repository/CharacterRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CharacterRepository extends CrudRepository<Character, Long> {
+public interface CharacterRepository extends CrudRepository<Character, Integer> {
 
     List<Character> findByName(String name);
